Move `DualTextBuffer` to root namespace - Move buffer to a better namespace
